requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
anyio==4.9.0
asgiref==3.8.1
asyncclick==8.1.8.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
frozenlist==1.6.2
idna==3.10
isodate==0.7.2
lxml==5.4.0
mashumaro==3.16
multidict==6.4.4
numpy==2.2.6
onvif_zeep==0.2.12
opencv-python==4.11.0.86
platformdirs==4.3.8
propcache==0.3.1
pycparser==2.22
pycryptodome==3.23.0
python-kasa==0.10.2
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rtp==0.0.4
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.0
zeep==4.3.1
